config/arm: split march cfg into arch and features
Older compilers may not support all arch versions and all features that
the target SoC supports, in which case it's better to figure out the
highest arch version and features that the compiler supports. Implement
a way to achieve this:
1. Find the highest arch version that the compiler supports, keeping in
mind the SoC arch version we're building. For example, if the SoC arch
version is arm8.2-a, but the compiler only supports arm8.1-a, use
arm8.1-a. On the other hand, if the compiler supports arm8.3-a (or
higher), use armv8.2-a.
2. With the architecture version locked, iterate over SoC features and
use all that are supported.

In all cases, emit a warning if there's something unsupported by the
compiler.

Signed-off-by: Juraj Linkeš <juraj.linkes@pantheon.tech>
Signed-off-by: 0-day Robot <robot@bytheb.org>