net/iavf: fix wrong FDIR L3 field set for IPv4 fragment packets
Originally, the value of field_selector for IPV4_FRAG header hdr1 is
the same as the previous header hdr2. For IPv4 packets, field_selector
for hdr2 can be any value between 0 and 4, depending on the selected
field. Actually, this value for IPV4_FRAG should be constant 0,
which denotes the field packet ID.

This patch adds an assignment to hdr1->field_selector to make sure that
it is always 0.

Fixes: 3334513ef484 ("net/iavf: support flow director for IP fragment")

Signed-off-by: Wenjun Wu <wenjun1.wu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>